okay ive tried to make an execute function but this code is a slow mess. Basically i need to finish and make the execute function actually work and fix the 29 bugs within it. Then i need to speed up the execute and probably refactor/mvoe alot of it around. But what i have done tried to skip past the fact that we were using the lookup and not reversing the bid prices, so i am attempting to convert to a negative so then we can keep the lookup array, which will help us with branch less programming in the the future hopefully. And then ofc change price from size_t to long so it can be a negative. But alot of bugs need to be worked out.

+++ include/order.hpp
@@ -46,7 +46,7 @@ class Order
 
     private:
         Order_type type;
-        size_t price;
+        long price;
         size_t size;
         size_t ID;
